moved manual mapping and overload code out of dinvoke
